Improvements and fixes at navigation
The method to obtain the target element is now a little more robust.

element.getBoundingClientRect() returns the position relativ to the view
port. But this works only as expected, if the view port is on the top of
the document.

If no internal target is given, the view port is set to top. This has to
be improved -- the tool should remember the last scroll positions.