Port to MacOS, step 1
Under MacOS, the application process can does still exist after
the last window was closed, which is not regarded at all yet.
After removing the first error messages, it occurred that the
IPC messaging between main and renderer process has to be
abstracted, because this is the only reason that all the modules
need the main window object.