Optimize app package size
Unfortunately, it is not yet possible to remove the superfluous locale
files, which take them self just under 30 MB. Choosing the locales to
include is currently only possible with macOS. A pre release of
electron-builder (tried 24.4.0) enabled this feature for other platforms
but this causes other problems, i.e. DevTools don't work anymore and
included documents that are linked in the README cannot be displayed.